week08/requirements.txt
# Week 8 Buổi 16 — Data Quality (Great Expectations)
great_expectations>=0.18.0,<0.19
pandas>=1.5.0
psycopg2-binary>=2.9.0
